Change game size type to u64

+++ native/hub/src/models/cloud_app.rs
@@ -16,14 +16,14 @@ pub struct CloudApp {
     #[serde(alias = "Last Updated")]
     pub last_updated: String,
     #[serde(alias = "Size (MB)", deserialize_with = "deserialize_size_mb_to_bytes")]
-    pub size: u32,
+    pub size: u64,
 }
 
-fn deserialize_size_mb_to_bytes<'de, D>(deserializer: D) -> Result<u32, D::Error>
+fn deserialize_size_mb_to_bytes<'de, D>(deserializer: D) -> Result<u64, D::Error>
 where
     D: serde::Deserializer<'de>,
 {
     let size_str = String::deserialize(deserializer)?;
     let size_mb = f64::from_str(&size_str).map_err(serde::de::Error::custom)?;
-    Ok((size_mb * 1000.0 * 1000.0) as u32)
+    Ok((size_mb * 1000.0 * 1000.0) as u64)
 }
